data-lab/job-intelligence/requirements.txt
playwright==1.58.0
pandas==3.0.0
python-dotenv==1.2.1
google-generativeai==0.8.6
tabulate==0.9.0
